requirements.txt
langchain
streamlit
openai
psycopg2-binary 
sqlalchemy
langchain-community
langchain-experimental
langgraph
python-dotenv
langchain-openai
langchain_tavily
bs4
langchain-openai
langgraph
streamlit
python-dotenv